Banking Application
we will learn how to create a mini-application for a banking system in Java. In this program, we will add some basic functionalities of a bank account like a deposit of amount, withdrawal of amount, etc.

Initially, the program accepts the number of customers we need to add and adds the customer and account details accordingly. Further, it displays the series of menus to operate over the accounts.

The series of menus displayed are as follows:

Display all account details
Search by account number
Deposit the amount
Withdraw the amount
Exit